feat: Move the Inline Scanner workload to the worker
